pyproject.toml
[project]
name = "legaltech-agentic-notice"
version = "0.1.0"
description = "Agentic workflow for complaint intake, policy/contact research, and legal notice drafting"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.40.0",
  "beautifulsoup4>=4.12.3",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  "reportlab>=4.1.0",
  "tldextract>=5.1.2",
  "uvicorn>=0.30.6"
]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
